+++ catalogo/catalogo/src/main/java/com/ecommerce/catalogo/domain/model/ItemCarrito.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -14,8 +15,8 @@ public class ItemCarrito {
     private Long carritoId;
     private Long productoId;
     private String nombreProducto;
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
     private Integer cantidad;
-    private Double subtotal;
+    private BigDecimal subtotal;
 
 }
