+++ catalogo/catalogo/src/main/java/com/ecommerce/catalogo/infraestructure/driver_adapters/jpa_repository/carrito/CarritoData.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -22,7 +23,7 @@ public class CarritoData {
 
     private Long usuarioId;
 
-    private Double precioTotal;
+    private BigDecimal precioTotal;
 
     @OneToMany(mappedBy = "carrito", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<ItemCarritoData> items;
